generics primitive type wrapper

+++ 15-generics-generics-and-primitive-types/src/main/java/org/pbansal/Main.java
@@ -12,9 +12,9 @@ public class Main {
 
      GenericList<Integer> numbers = new GenericList<>();
      numbers.add(1); // boxing
-     int number = (int) numbers.get(0);
+     int number = numbers.get(0); // unboxing
         System.out.println(number);
-
 
     }
 }
+
